Add test for start with jitter
